backend/pyproject.toml
[project]
name = "ai-in-aec-kg-backend"
version = "0.1.0"
description = "Backend for AI in AEC Knowledge Graph"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "neo4j>=5.28.0",
    "openai>=1.82.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "python-jose[cryptography]>=3.4.0",
    "aiosqlite>=0.21.0",
    "rapidfuzz>=3.13.0",
    "spacy>=3.8.0",
    "keybert>=0.8.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.26.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
